AG-9914 Resolve circular dependency
